ARM OMAP2+ OPP: Set clock rate for missing devices
ARM OMAP2+ OPP: Set clock rate for missing devices

Currently OPP layer simply skips the OPP for devices which are
not available. This might lead invalid OPP configuration. So
device clocks for all those devices that not available should
be set to minimum rate so that system will always be in sane
OPP configuration.

Change-Id: I30938a4aae89b6fc5057d467a0f9edcb33d2d2bc
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>
[ivan.khoronzhuk@ti.com: move set_device_opp to __init]
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>
[grygorii.strashko@ti.com: move omap_opp_set_min_rate to __init]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>