PMIC: Do not hardcode I2C bus contents
By looking for specific entries in the I2C board data arrays, the
routine omap_pmic_late_init() effectively hard codes what is on the
I2C busses.  The values used are specific to the TI reference platforms
and do not apply to other OMAP4 boards like the Phytec PCM049.

Signed-off-by: Steve Schefter <steve@scheftech.com>