OMAP4: HDMI: Fix bugs
The use of GPIOs for ct_cp_hpd and ls_oe is specific to TI
development boards such as Blaze, but not other boards such as the
Phytec phyCORE-OMAP44xx.  The HDMI panel probe routine should therefore
not fail if these GPIOs are not defined.

The routine hdmi.hdmi_power_on_cb function pointer is only
initialized on a system compiled with OMAP5 support.  The driver
must therefore check that it is set before trying to call the
routine to avoid a panic in OMAP4-only kernels.

Signed-off-by: Steve Schefter <steve@scheftech.com>