OMAP: Ignore undocumented pending interrupt
When entering suspend, the routine omap_wakeupgen_check_interrupts()
checks for pending interrupts.  If there are pending interrupts,
suspend is aborted.  The idea is that if an interrupt is about to
occur, we'd immediately wake up anyway.  However, one interrupt,
numer 40, is always pending.  While this interrupt is undocumented
in the TRM, it is #defined in the code as being a security
interrupt.  Ignore this interrupt when checking for pending
interrupts.

Signed-off-by: Steve Schefter <steve@scheftech.com>