OMAP PWM: Do not attempt to configure a disabled PWM
The generic backlight PWM driver (drivers/video/backlight/pwm_bl.c)
does not keep track of the state.  This resulted in the following
steps leading to an L3 crash:  The display is turned off in early
suspend, leading to backlight_store_update_status with a brightness
of 0.  The backlight PWM driver disables the PWM with brightness
0 (pwm_disable in arch/arm/plat-omap/pwm.c).  This causes the timer
to be stopped and then its hwmod being disabled.  Later, the
system enters suspend and pwm_backlight_suspend() calls
pwm_config() with a brightness of 0.  pwm_config() attempts to
access registers of a hardware module that does not have its
interface clocks enabled.

Since the OMAP PWM driver already keeps track of the state, use
that to decide if a call to pwm_config() is valid or not before
accessing any registers.