ksp5012: support added for fpga version check
The i2c initialization for RTC, as well as the entire
libs_pwr/battery_libs initialization has been offloaded
to

	arch/arm/mach-omap2/board-omap4ksp5012-version.c

which can only be built as a module. This module contains
a module param bus_id, which allows the user to select
which i2c bus number to register the RTC and libs_pwr
devices on.

If loading it manually, it has only been tested with bus_id
1, 3, and 4. It does not provide complete error checking for
other bus numbers, and due to the nature of its hooks into the
i2c controller, it cannot be removed using 'rmmod' at this time.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>