Add support for the LTC1760 battery charger with two ND2054 batteries.
This driver is not fully debugged at this time due to hardware problems.
I2C bus access to the batteries is not currently working.  The
batteries' I2C acks appear to the OMAP4 as nacks.

In particular, the calls to power_supply_register() may need some
adjustment.  These were moved from nd2054_battery_probe() to
nd2054_update() and called only when the batteries are actually
found.  This is necessary since Android will power off the device if
the driver repeatedly returns a failure in its get_property routine.
This repeated failure would be the case if the batteries are
removed (and is the case now with hardware problems preventing
access to the batteries).