net: wireless: wl12xx: fix broken dependency on Kconfig macros
When building for other MACH types, the other
CONFIG_MACH_* definitions will not be set so
the driver is unaware of their existence.
Instead of using them directly in a runtime
check, switch to using an integer that gets set
during compilation and is then defined 100% of
the time during operation.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>